+++ src/drawHistory.js
@@ -1,0 +1,6 @@
+export function drawHistory(elTwo, arr) {
+  elTwo.innerHTML = `HISTORY:${arr
+    .map((item) => `<span style="margin: 2px;">${item}</span>`)
+    .reverse()
+    .join("")}`;
+}
