+++ src/getWeather.js
@@ -1,0 +1,6 @@
+export async function getWeather(inputValue) {
+  const apiKey = "92dae48bac98a7191c6227716a76ac12";
+  const url = `https://api.openweathermap.org/data/2.5/forecast?q=${inputValue}&units=metric&appid=${apiKey}`;
+  const response = await fetch(url);
+  return response.json();
+}
