+++ src/readHistory.js
@@ -1,0 +1,4 @@
+export async function readHistory() {
+  const history = localStorage.getItem("history");
+  return JSON.parse(history) ?? [];
+}
